Extract partial classes for `DistributionExtensions`
